move nerium init to main module
